requirements.txt
astroid==2.4.2
Django==2.1.5
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.7
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
